requirements.txt
matplotlib
numpy
pandas
pillow
streamlit
yfinance